Extract activities outside internal folder.
