+++ server/neptune/workflows/activities/execute.go
@@ -7,7 +7,7 @@ import (
 	"os/exec"
 
 	"github.com/pkg/errors"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/execute"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/execute"
 )
 
 type executeCommandActivities struct{}
