+++ server/neptune/workflows/activities/github/link/builder.go
@@ -5,8 +5,8 @@ import (
 	"net/url"
 	"path"
 
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/terraform"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/terraform"
 )
 
 type Builder struct{}
