+++ server/neptune/workflows/activities/terraform/root.go
@@ -1,8 +1,8 @@
 package terraform
 
 import (
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/execute"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/execute"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github"
 )
 
 // Root is the definition of a root
