+++ server/neptune/workflows/internal/deploy/terraform/deployment.go
@@ -3,8 +3,8 @@ package terraform
 import (
 	"fmt"
 
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/terraform"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/terraform"
 
 	"github.com/google/uuid"
 )
