+++ server/neptune/workflows/internal/terraform/root_fetcher.go
@@ -2,8 +2,8 @@ package terraform
 
 import (
 	"github.com/pkg/errors"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/activities/terraform"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/terraform"
 	"go.temporal.io/sdk/workflow"
 )
